front/requirements.txt
solara==1.19.0  # Pip freeze
python-dotenv==1.0.0
pydantic==1.10.9
bson==0.5.10
requests==2.31.0
pandas==2.0.3
paginate==0.5.6
PyJWT==2.8.0
